feat: add rotas ao projeto, estruturando as pastas
