--- App.tsx
@@ -1,6 +1,0 @@
-import Home from "./src/pages/home/home";
-// import Login from "./src/pages/login";
-
-export default function App() {
-  return <Home />;
-}
