+++ app.json
@@ -4,11 +4,11 @@
     "slug": "CorraAgil",
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/icon.png",
+    "icon": "../assets/icon.png",
     "scheme": "CorraAgil",
     "userInterfaceStyle": "light",
     "splash": {
-      "image": "./assets/splash.png",
+      "image": "../assets/splash.png",
       "resizeMode": "contain",
       "backgroundColor": "#12263A"
     },
@@ -17,12 +17,12 @@
     },
     "android": {
       "adaptiveIcon": {
-        "foregroundImage": "./assets/icon.png",
+        "foregroundImage": "../assets/icon.png",
         "backgroundColor": "#12263A"
       }
     },
     "web": {
-      "favicon": "./assets/icon.png"
+      "favicon": "../assets/icon.png"
     },
     "plugins": ["expo-router"]
   }
