--- src/pages/login/styles.ts
@@ -1,33 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#12263A",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  top: {
-    alignItems: "center",
-    marginBottom: 40,
-  },
-  title: {
-    color: "#FFF",
-    fontWeight: "bold",
-    fontSize: 32,
-  },
-  containerInput: {
-    marginTop: 30,
-    alignItems: "center",
-  },
-
-  forgetPassword: {
-    color: "#FFF",
-    fontWeight: "bold",
-    fontSize: 20,
-    marginBottom: 33,
-  },
-  containerButton: {
-    marginBottom: 96,
-  },
-});
